@@ -1,68 +1,80 @@
-import React, { FC, useState, useRef, Fragment } from 'react';
-import clsx from 'clsx';
 import { RecordOf } from 'immutable';
-import Alert from '@material-ui/lab/Alert';
-import { redirect } from '@graasp/sdk';
-import { makeStyles } from '@material-ui/core/styles';
-import { getEmbeddedLinkExtra } from '../utils/itemExtra';
-import OpenInNewIcon from '@material-ui/icons/OpenInNew';
-import Button from '../Button';
-import withCaption from './withCaption';
-import { ITEM_MAX_HEIGHT } from '../constants';
-import type { EmbeddedLinkItemExtra, Item } from '../types';
-import withResizing from './withResizing';
 
-interface LinkItemProps {
-  item: RecordOf<Item<EmbeddedLinkItemExtra>>;
-  height?: number | string;
-  onSaveCaption?: (text: string) => void;
+import OpenInNewIcon from '@mui/icons-material/OpenInNew';
+import { styled } from '@mui/material';
+import Alert from '@mui/material/Alert';
+
+import React, { FC, Fragment, useRef, useState } from 'react';
+
+import { Item, redirect } from '@graasp/sdk';
+
+import Button from '../buttons/Button';
+import { ITEM_MAX_HEIGHT } from '../constants';
+import type { EmbeddedLinkItemExtra, MemberRecord } from '../types';
+import { getEmbeddedLinkExtra } from '../utils/itemExtra';
+import withCaption from './withCaption';
+import withResizing, { StyledIFrame } from './withResizing';
+
+export interface LinkItemProps {
+  member?: MemberRecord;
   editCaption?: boolean;
-  showCaption?: boolean;
-  saveButtonId?: string;
-  loadingMessage?: string;
-  openLinkMessage?: string;
   errorMessage?: string;
+  height?: number | string;
+  /**
+   * whether the link can be resized
+   */
   isResizable?: boolean;
+  item: RecordOf<Item<EmbeddedLinkItemExtra>>;
+  loadingMessage?: string;
+  onSaveCaption?: (text: string) => void;
+  openLinkMessage?: string;
+  /**
+   * id of the save button
+   */
+  saveButtonId?: string;
+  /**
+   * whether the caption should be displayed
+   */
+  showCaption?: boolean;
+  /**
+   * whether the iframe should be displayed
+   */
+  showIframe?: boolean;
+  /**
+   * whether the button should be displayed
+   */
+  showButton?: boolean;
 }
 
-const useStyles = makeStyles((theme) => ({
-  iframe: {
-    width: '100%',
-    border: 'none',
-  },
-  iframeWithoutResizer: {
-    maxHeight: ITEM_MAX_HEIGHT,
-    height: ITEM_MAX_HEIGHT,
-  },
-  linkButton: {
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    marginTop: theme.spacing(1),
-  },
-  iframeContainer: {
-    position: 'relative',
-    overflow: 'auto',
-  },
-  iframeContainerWithoutResizer: {
-    maxHeight: ITEM_MAX_HEIGHT,
-  },
+const IFrameContainer = styled('div')({
+  position: 'relative',
+  maxHeight: ITEM_MAX_HEIGHT,
+  overflow: 'auto',
+});
+
+const StyledLinkButton = styled(Button)(({ theme }) => ({
+  marginLeft: 'auto',
+  marginRight: 'auto',
+  marginTop: theme.spacing(1),
 }));
 
 const LinkItem: FC<LinkItemProps> = ({
   item,
+  member,
   onSaveCaption,
   saveButtonId,
   editCaption = false,
   showCaption = true,
+  showIframe = false,
+  showButton = true,
   loadingMessage = 'Link is Loading...',
   openLinkMessage = 'Click here to open the link manually',
-  height: defaultHeight,
+  height: defaultHeight = 400,
   errorMessage = 'The link is malformed.',
   isResizable = false,
 }) => {
-  const classes = useStyles();
   const [isLoading, setIsLoading] = useState(true);
-  const [height] = useState<string | number>(defaultHeight ?? '100%');
+  const [height] = useState<string | number>(defaultHeight);
   const iframeRef = useRef<HTMLIFrameElement>(null);
 
   const id = item.id;
@@ -78,7 +90,6 @@ const LinkItem: FC<LinkItemProps> = ({
   // if available, display specific player
   const html = extra?.html;
   if (html) {
-    // eslint-disable-next-line react/no-danger
     const component = (
       <div id={id} dangerouslySetInnerHTML={{ __html: html }} />
     );
@@ -108,54 +119,58 @@ const LinkItem: FC<LinkItemProps> = ({
     redirect(url, { openInNewTab: true });
   };
 
-  const iframe = (
-    <iframe
-      id={id}
-      className={clsx(classes.iframe, {
-        [classes.iframeWithoutResizer]: !isResizable,
-      })}
-      title={name}
-      src={url}
-      onLoad={handleLoad}
-      height='100%'
-      ref={iframeRef}
-    />
+  const renderIframe = (): JSX.Element | null => {
+    if (!showIframe) {
+      return null;
+    }
+
+    const iframe = (
+      <StyledIFrame
+        height={height}
+        id={id}
+        isResizable={isResizable}
+        onLoad={handleLoad}
+        ref={iframeRef}
+        src={url}
+        title={name}
+      />
+    );
+
+    const ResizableLink = withResizing({
+      height,
+      component: iframe,
+      memberId: member?.id,
+      itemId: item.id,
+    });
+
+    return (
+      <>
+        <IFrameContainer hidden={!isLoading} style={{ height }}>
+          {loadingMessage}
+        </IFrameContainer>
+        <div hidden={isLoading}>
+          {isResizable ? (
+            <div>
+              <ResizableLink />
+            </div>
+          ) : (
+            iframe
+          )}
+        </div>
+      </>
+    );
+  };
+
+  const button = (
+    <StyledLinkButton onClick={onClick} startIcon={<OpenInNewIcon />}>
+      {item.name ?? openLinkMessage}
+    </StyledLinkButton>
   );
 
   const component = (
     <Fragment>
-      <div
-        hidden={!isLoading}
-        className={classes.iframeContainer}
-        style={{ height: height }}
-      >
-        {loadingMessage}
-      </div>
-      <div
-        hidden={isLoading}
-        className={clsx(classes.iframeContainer, {
-          [classes.iframeContainerWithoutResizer]: !isResizable,
-        })}
-      >
-        {isResizable ? (
-          <div>
-            {withResizing({
-              height,
-            })(iframe)}
-          </div>
-        ) : (
-          iframe
-        )}
-      </div>
-      {isLoading && (
-        <Button
-          onClick={onClick}
-          className={classes.linkButton}
-          startIcon={<OpenInNewIcon />}
-        >
-          {openLinkMessage}
-        </Button>
-      )}
+      {renderIframe()}
+      {(isLoading || showButton) && button}
     </Fragment>
   );
 
